apps/backend/pyproject.toml
[project]
name = "finance-os-backend"
version = "0.1.0"
description = "Production-minded Django backend for a personal finance operating system."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<5.3",
  "djangorestframework>=3.15,<3.17",
  "django-environ>=0.12,<0.13",
  "django-cors-headers>=4.7,<4.8",
  "psycopg[binary]>=3.2,<3.4",
  "PyJWT[crypto]>=2.10,<3.0",
  "httpx>=0.28,<0.29",
  "drf-spectacular>=0.28,<0.29",
  "python-dateutil>=2.9,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-django>=4.11,<5.0",
  "factory-boy>=3.3,<4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["apps*", "api*", "config*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
pythonpath = ["."]
testpaths = ["tests"]
